+++ changelog.md
@@ -1,6 +1,9 @@
 # changelog
 
 ## 3.2.4
+* `NEW` settings:
+  + `workspace.supportScheme`: `["file", "untitled", "git"]`
+  + `diagnostics.disableScheme`: `["git"]`
 * `FIX` hover: can not union `table` with other basic types
 * `FIX` [#1125](https://github.com/sumneko/lua-language-server/issues/1125)
 * `FIX` [#1131](https://github.com/sumneko/lua-language-server/issues/1131)